+++ example/F0.Wpf.Example/F0.Wpf.Example.Windows.Data/App.xaml.cs
@@ -1,0 +1,8 @@
+﻿using System.Windows;
+
+namespace F0.Wpf.Example.Windows.Data
+{
+	internal partial class App : Application
+	{
+	}
+}