+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Added `Windows.Data.CompositeValueConverter` chaining multiple `System.Windows.Data.IValueConverter`.
 
 ## v0.1.0 (2018-09-18)
-- Added Diagnostics.DataBindingTraceListener detecting data-binding errors.
+- Added `Diagnostics.DataBindingTraceListener` detecting data-binding errors.